+++ src/ds_hir.h
@@ -148,6 +148,7 @@ typedef struct {
 typedef struct {
     DsStr name;
     bool has_default;
+    DsLowerValueKind default_kind;
     DsLowerExpr *default_value;
     DsSpan span;
 } DsLowerFnParam;