+++ src/vm_internal.h
@@ -30,6 +30,7 @@ typedef enum {
     OP_GET_INDEX,
     OP_PUSH_ARRAY,
     OP_FOR_ARRAY,
+    OP_RESET_FOR,
     OP_ASSERT,
     OP_RETURN_FUNC,
     OP_RETURN,